chore: 함수 각각 파일로 분리
